old-tests/submission/Infraware/Forms/contents/Forms/textarea_willvalidate.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_willValidate</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
		<p><textarea id='textarea_id'></textarea></p>
	</form>
	<script>

		var textarea = document.getElementById("textarea_id");

		if (typeof(textarea.willValidate) == "boolean") {
			test( function() {
				assert_equals(textarea.willValidate, true, "willValidate attribute is not correct.");
			});
		} else {
			test( function() {
				assert_unreached("willValidate attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
